updated rainfall timesereis
